Consolidate run and test logic.
